+++ pages/_app.js
@@ -3,6 +3,7 @@ import '../styles/globals.css';
 import { useRouter } from 'next/router';
 import { createTheme, ThemeProvider } from '@mui/material';
 import { UserProvider } from '@auth0/nextjs-auth0';
+import { ContextWrapper } from '../src/AppContext';
 
 function MyApp({ Component, pageProps }) {
 	const { route } = useRouter();
@@ -35,7 +36,9 @@ function MyApp({ Component, pageProps }) {
 	return (
 		<UserProvider>
 			<ThemeProvider theme={theme}>
-				<Component {...pageProps} />
+				<ContextWrapper>
+					<Component {...pageProps} />
+				</ContextWrapper>
 			</ThemeProvider>
 		</UserProvider>
 	);
